+++ src/main/java/ru/itis/inf304/ConstructionLab12Sem2/WorkClasses/WallAndCeilingDecoration.java
@@ -3,7 +3,7 @@ package ru.itis.inf304.ConstructionLab12Sem2.WorkClasses;
 import ru.itis.inf304.ConstructionLab12Sem2.AbstractWorkClass;
 
 public class WallAndCeilingDecoration extends AbstractWorkClass {
-    long duration = 7;
+    int duration = 7;
     int dayOfBegin;
     int dayOfEnd;
 
@@ -29,7 +29,7 @@ public class WallAndCeilingDecoration extends AbstractWorkClass {
     }
 
     @Override
-    protected long getDuration() {
+    protected int getDuration() {
         return duration;
     }
